move module code to lib.rs
